[Refactor] W7 피드백(리턴 타입 맞춤)(#34)
